first cut
