+++ common/errors/shutdownerrors.go
@@ -1,0 +1,53 @@
+// The MIT License (MIT)
+
+// Copyright (c) 2017-2020 Uber Technologies Inc.
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+package errors
+
+import (
+	"errors"
+	"time"
+)
+
+var ErrProcessorShutdown = errors.New("queue processor has been shutdown")
+
+type ErrShardClosed struct {
+	Msg      string
+	ClosedAt time.Time
+}
+
+var _ error = (*ErrShardClosed)(nil)
+
+func (e *ErrShardClosed) Error() string {
+	return e.Msg
+}
+
+// a shutdown error is a type of nonretriable error for task processors
+func IsShutdownError(err error) bool {
+	var errShardClosed *ErrShardClosed
+	if errors.As(err, &errShardClosed) {
+		return true
+	}
+	if errors.Is(err, ErrProcessorShutdown) {
+		return true
+	}
+	return false
+}
