+++ service/history/queue/timer_queue_processor.go
@@ -22,7 +22,6 @@ package queue
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sync"
 	"sync/atomic"
@@ -30,6 +29,7 @@ import (
 
 	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/clock"
+	cerrors "github.com/uber/cadence/common/errors"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
@@ -281,7 +281,7 @@ func (t *timerQueueProcessor) FailoverDomain(domainIDs map[string]struct{}) {
 	actionResult, err := t.HandleAction(context.Background(), t.currentClusterName, NewGetStateAction())
 	if err != nil {
 		t.logger.Error("Timer failover failed while getting queue states", tag.WorkflowDomainIDs(domainIDs), tag.Error(err))
-		if err == errProcessorShutdown {
+		if err == cerrors.ErrProcessorShutdown {
 			// processor/shard already shutdown, we don't need to create failover queue processor
 			return
 		}
@@ -362,14 +362,14 @@ func (t *timerQueueProcessor) HandleAction(ctx context.Context, clusterName stri
 		if ctxErr := ctx.Err(); ctxErr != nil {
 			return nil, ctxErr
 		}
-		return nil, errProcessorShutdown
+		return nil, cerrors.ErrProcessorShutdown
 	}
 
 	select {
 	case resultNotification := <-resultNotificationCh:
 		return resultNotification.result, resultNotification.err
 	case <-t.shutdownChan:
-		return nil, errProcessorShutdown
+		return nil, cerrors.ErrProcessorShutdown
 	case <-ctx.Done():
 		return nil, ctx.Err()
 	}
@@ -425,8 +425,7 @@ func (t *timerQueueProcessor) completeTimerLoop() {
 				}
 
 				t.logger.Error("Failed to complete timer task", tag.Error(err))
-				var errShardClosed *shard.ErrShardClosed
-				if errors.As(err, &errShardClosed) {
+				if cerrors.IsShutdownError(err) {
 					if !t.shard.GetConfig().QueueProcessorEnableGracefulSyncShutdown() {
 						go t.Stop()
 						return
